08 07 2023/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.2.0
CTkMessagebox==2.0
customtkinter==5.2.0
darkdetect==0.8.0
dataclasses-json==0.5.8
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.1
fakeyou==1.1.1
filelock==3.12.2
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.220
langchainplus-sdk==0.0.17
llvmlite==0.40.1
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
notion-client==2.0.0
numba==0.57.1
numexpr==2.8.4
numpy==1.24.4
openai==0.27.8
openai-whisper @ git+https://github.com/openai/whisper.git@f572f2161ba831bae131364c3bffdead7af6d210
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
PyAudio==0.2.13
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.10.10
pydub==0.25.1
pygame==2.5.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dotenv==1.0.0
pyttsx3==2.90
pytweening==1.0.7
pywhatkit==5.4
pywin32==306
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.17
sympy==1.12
tenacity==8.2.2
tiktoken==0.3.3
torch==2.0.1
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
typing-extensions==4.7.0
typing-inspect==0.9.0
urllib3==2.0.3
Werkzeug==2.3.6
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
